Fixing the issues
